requirements.txt
# LangChain & local LLM (Ollama + Mistral)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Web scraping & search
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Config
python-dotenv>=1.0.0
